string prompt template
